Data reduction in 1 script reduce_data.py for both blue and red arms. Implemented the data classification in data_classifier.py, previously done in generate_metadata.py. Also creates the directories for reduced data. Built 2 .json templates with the DRP setup parameters. Updated steps.py accordingly and .gitignore. Removed a few defunct scripts.
