+++ .gitignore
@@ -55,3 +55,13 @@ docs/_build/
 
 # PyBuilder
 target/
+
+# MacOS files
+.DS_Store
+reduction_scripts/.DS_Store
+reduction_scripts/extra_tools/.DS_Store
+
+# Directories to test the DRP
+reduction_scripts/raw_data/
+reduction_scripts/reduc_blue/
+reduction_scripts/reduc_red/
