+++ reduction_scripts/blue_params.json
@@ -1,0 +1,205 @@
+[
+    {
+        "step": "overscan_sub",
+        "run": true,
+        "suffix": "00",
+        "args": {}
+    },
+    {
+        "step": "bpm_repair",
+        "run": true,
+        "suffix": "01",
+        "args": {}
+    },
+    {
+        "step": "superbias",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "method": "row_med",
+            "plot": false,
+            "verbose": false
+        }
+    },
+    {
+        "step": "bias_sub",
+        "run": true,
+        "suffix": "02",
+        "args": {
+            "method": "subtract",
+            "plot": false,
+            "verbose": false
+        }
+    },
+    {
+        "step": "superflat",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "source": "dome"
+        }
+    },
+    {
+        "step": "superflat",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "source": "twi",
+            "scale": "median_nonzero"
+        }
+    },
+    {
+        "step": "slitlet_profile",
+        "run": true,
+        "suffix": null,
+        "args": {}
+    },
+    {
+        "step": "flat_cleanup",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "type": [
+                "dome"
+            ],
+            "verbose": true,
+            "plot": false,
+            "offsets": [
+                0.4,
+                0.4
+            ],
+            "radius": 10.0,
+            "nsig_lim": 3.0
+        }
+    },
+    {
+        "step": "superflat_mef",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "source": "dome"
+        }
+    },
+    {
+        "step": "superflat_mef",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "source": "twi"
+        }
+    },
+    {
+        "step": "slitlet_mef",
+        "run": true,
+        "suffix": "03",
+        "args": {
+            "ns": true
+        }
+    },
+    {
+        "step": "wave_soln",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "verbose": true,
+            "method": "optical",
+            "shift_method": "xcorr_all",
+            "find_method": "mpfit",
+            "doalphapfit": true,
+            "doplot": false,
+            "dlam_cut_start": 5.0,
+            "multithread": false
+        }
+    },
+    {
+        "step": "wire_soln",
+        "run": true,
+        "suffix": null,
+        "args": {}
+    },
+    {
+        "step": "flat_response",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "mode": "all"
+        }
+    },
+    {
+        "step": "cosmic_rays",
+        "run": true,
+        "suffix": "04",
+        "args": {
+            "ns": true,
+            "multithread": false
+        }
+    },
+    {
+        "step": "sky_sub",
+        "run": true,
+        "suffix": "05",
+        "args": {
+            "ns": true
+        }
+    },
+    {
+        "step": "obs_coadd",
+        "run": true,
+        "suffix": "06",
+        "args": {
+            "method": "sum"
+        }
+    },
+    {
+        "step": "flatfield",
+        "run": true,
+        "suffix": "07",
+        "args": {}
+    },
+    {
+        "step": "cube_gen",
+        "run": true,
+        "suffix": "08",
+        "args": {
+            "multithread": false,
+            "adr": true,
+            "wmin_set": 3500.0,
+            "wmax_set": 5700.0
+        }
+    },
+    {
+        "step": "extract_stars",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "ytrim": 4,
+            "type": "flux"
+        }
+    },
+    {
+        "step": "derive_calib",
+        "run": true,
+        "suffix": null,
+        "args": {
+            "plot_stars": false,
+            "plot_sensf": false,
+            "polydeg": 25,
+            "excise_cut": 0.005,
+            "method": "poly",
+            "boxcar": 10,
+            "norm_stars": true
+        }
+    },
+    {
+        "step": "flux_calib",
+        "run": true,
+        "suffix": "10",
+        "args": {}
+    },
+    {
+        "step": "save_3dcube",
+        "run": true,
+        "suffix": "11",
+        "args": {}
+    }
+]
